+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Show DelphiLint actions in IDE Insight search bar
 * Allow DelphiLint actions to be added to editor toolbars
 * Keyboard shortcuts for Analyze This File (Shift+Ctrl+L) and Analyze All Open Files (Shift+Ctrl+Alt+L)
+* Friendly error messages for common analysis errors
 
 ## [0.1.0] - 2023-07-20